AWS EC2 Tagger Plugin and ServiceProcessor Interface
Tags metrics with EC2 Metadata and EC2 Instance Tags.
This also introduce a new ServiceProcessor interface similar to the existing ServiceInput interface.

The plugin would fetch the EC2 Metadata and EC2 Instance Tags at the start, and regularly update the EC2 Instance Tags.
These tags can be configured to be added to the metrics.