fix: Add source field to keep the original mount source if its symlink
Under linux, the disk.Partitions method would evaluate the symbolic link
of the mount source is under /dev/mapper, which could be useful
information under some cases. In this change, we add a new field named
source and would only be populated when the symlink evaluation had been
carried out.

The json tag of the new source field is created with "omitempty" option,
so it won't create any behavior difference for code that does not make
use of it.